Cleaned up widgetRepo

+++ BAR.DAL/WidgetRepository.cs
@@ -58,7 +58,8 @@ namespace BAR.DAL
 		public Dashboard ReadDashboardWithWidgets(int dashboardId)
 		{
 			return ctx.Dashboards.Include(dash => dash.Widgets)
-				.Where(dash => dash.DashboardId == dashboardId).SingleOrDefault();
+								 .Where(dash => dash.DashboardId == dashboardId)
+								 .SingleOrDefault();
 		}
 
 		/// <summary>
@@ -68,17 +69,18 @@ namespace BAR.DAL
 		public Dashboard ReadDashboardWithWidgets(string userId)
 		{
 			return ctx.Dashboards.Include(dash => dash.Widgets)
-				.Where(dash => dash.User.Id == userId).FirstOrDefault();
+								 .Where(dash => dash.User.Id == userId)
+								 .SingleOrDefault();
 		}
 
 		/// <summary>
 		/// Gives back the general dashboard.
-		/// 
 		/// </summary>
 		public Dashboard ReadGeneralDashboard()
 		{
 			return ctx.Dashboards.Include(dash => dash.Widgets)
-				.Where(dash => dash.DashboardType == DashboardType.General).FirstOrDefault();
+								 .Where(dash => dash.DashboardType == DashboardType.General)
+								 .SingleOrDefault();
 		}
 
 		/// <summary>
@@ -191,7 +193,7 @@ namespace BAR.DAL
 			foreach (int id in dashboardIds)
 			{
 				Dashboard dashboardToDelete = ReadDashboardWithWidgets(id);
-				ctx.Dashboards.Remove(dashboardToDelete);
+				if (dashboardToDelete != null) ctx.Dashboards.Remove(dashboardToDelete);
 			}
 			return ctx.SaveChanges();
 		}
@@ -213,7 +215,8 @@ namespace BAR.DAL
 		{
 			return ctx.Widgets.Include(widget => widget.PropertyTags)
 							  .Include(widget => widget.Items)
-							  .Where(widget => widget.WidgetId == widgetid).SingleOrDefault();
+							  .Where(widget => widget.WidgetId == widgetid)
+							  .SingleOrDefault();
 		}
 
 		/// <summary>
@@ -221,8 +224,7 @@ namespace BAR.DAL
 		/// </summary>
 		public IEnumerable<Widget> ReadAllWidgetsWithAllItems()
 		{
-			return ctx.Widgets.Include(Widget => Widget.PropertyTags)
-							  .Include(widget => widget.Items).AsEnumerable();
+			return ctx.Widgets.Include(widget => widget.Items).AsEnumerable();
 		}
 
 		/// <summary>
@@ -285,9 +287,9 @@ namespace BAR.DAL
 		public IEnumerable<Widget> ReadAllWidgetsWithAllDataForItem(int itemId)
 		{
 			Item itemToReturn =  ctx.Items.Include(item => item.ItemWidgets)
-							.Include(item => item.ItemWidgets.Select(widget => widget.WidgetDatas))
-							.Include(item => item.ItemWidgets.Select(widget => widget.WidgetDatas.Select(data => data.GraphValues)))
-							.Where(item => item.ItemId == itemId).SingleOrDefault();
+										  .Include(item => item.ItemWidgets.Select(widget => widget.WidgetDatas))
+									      .Include(item => item.ItemWidgets.Select(widget => widget.WidgetDatas.Select(data => data.GraphValues)))
+										  .Where(item => item.ItemId == itemId).SingleOrDefault();
 			return itemToReturn.ItemWidgets.AsEnumerable();
 		}
 
@@ -313,7 +315,6 @@ namespace BAR.DAL
 								  .Include(data => data.GraphValues)
 								  .AsEnumerable();
 		}
-
 
 		/// <summary>
 		/// Creates a range of WidgetData.
